Update training command for concept2clip to include gan w space training
